Create tiny_tokenizer.word_tokenizers
