+++ tiny_tokenizer/__init__.py
@@ -1,5 +1,6 @@
 """__init__.py."""
 from .sentence_tokenizer import SentenceTokenizer  # NOQA
 from .word_tokenizer import WordTokenizer  # NOQA
+from .token import Token
 
 __version__ = "2.1.0"
